feat(dsg) rename file update imports and refactor get entity id type utility
