+++ packages/amplication-data-service-generator/src/util/get-entity-id-type.ts
@@ -3,17 +3,17 @@ import DsgContext from "../dsg-context";
 import { EnumDataType } from "../models";
 import { USER_ENTITY_NAME } from "../server/user-entity";
 
-export const getUserIdType = () => {
-  const { entities } = DsgContext.getInstance;
-  const userEntity = entities.find(
-    (entity) => entity.name === USER_ENTITY_NAME
-  );
+export const getUserIdType = () => getEntityIdType(USER_ENTITY_NAME);
 
-  if (!userEntity) {
+export const getEntityIdType = (entityName: string) => {
+  const { entities } = DsgContext.getInstance;
+  const entity = entities.find((entity) => entity.name === entityName);
+
+  if (!entity) {
     throw new Error("User entity not found");
   }
 
-  const idField = userEntity.fields.find(
+  const idField = entity.fields.find(
     (field) => field.dataType === EnumDataType.Id
   );
   if (!idField) {
